Addresses new CRAN warning regarding using sprintf
A new CRAN warning as of 2023-02-13 17:54:56 CET discorages the use of
sprintf, probably because of the potential for a buffer overflow.

Result: WARN
    File ‘PCICt/libs/PCICt.so’:
     Found ‘__sprintf_chk’, possibly from ‘sprintf’ (C)
     Object: ‘datetime_360.o’

    Compiled code should not call entry points which might terminate R nor
    write to stdout/stderr instead of to the console, nor use Fortran I/O
    nor system RNGs nor [v]sprintf.

A very close read of the code indicates that triggering a buffer
overflow in this code would be impossible, as the write length is
constrained (to 6 bytes). However, the fix to this warning is simple.

This patch simply replaces sprintf with snprintf.