Pass format parameters through for 360 day cals
All of the PCICt classes methods accept ... parameters for printing
and formating, but apparently they are not passed through in the case
of 360 day calendars. This patch corrects that by passing them through
to R's native format method (after all of the custom date formatting
is complete), and adds a few test cases.

Fixes #2